Moved buffer to convenience library

+++ LogicGen.Compute/Shaders/BufferDescription.cs
@@ -2,14 +2,14 @@ namespace LogicGen.Compute.Shaders;
 
 public interface IBufferDescription {
 	public uint BindingIndex { get; }
-	public uint Size { get; }
+	public ulong Size { get; }
 }
 
 public class BufferDescription : IBufferDescription {
-	public BufferDescription(uint bindingIndex, uint size) {
+	public BufferDescription(uint bindingIndex, ulong size) {
 		BindingIndex = bindingIndex;
 		Size = size;
 	}
 	public uint BindingIndex { get; }
-	public uint Size { get; }
+	public ulong Size { get; }
 }
